Changes in testcases
